# feat(eventListener): fix #798, improve EventTarget.addEventListener performance