fix #798.

1. improve `EventTarget.addEventListener` performance with @mhevery 's idea which described in #798, we will not save a `eventTask` list on `event target`, and will not go through this `eventTask` list to find whether we have already registered the event or not. 

instead we save the information to `callback` itself, so 

```
div.addEventListener('click', () => {}); 
```

will not slow down

```
div.addEventListener('mouseover', () => {}); 
```

2. this solution only work in browser, in `nodejs` we still use the original way to save `taskList`, 
because in the `eventListener` callback, there is no `eventName` information, and `nodejs` support to add the identical listener multiple times.