@mhevery 2017-06-12
I added a performance test here: https://github.com/angular/zone.js/commit/3a017d3fb10a9bcd6fb08e9e9527b9c5d0d53402
Just run `gulp build` and than open the file in the browser to see how fast it is. In DevTools go to `Performance` tab and record the click.  Do `bottom-up` view to see what is going on.

<img width="584" alt="screen shot 2017-06-12 at 2 53 40 am" src="https://user-images.githubusercontent.com/111951/27015983-98590f48-4f1a-11e7-964f-bc8ee0d926d0.png">

Right away two things jump out. 
1. `Function.toString` should not be called (I left a comment where it is coming from.)
2. I see try-catch which needs to go (left a comment)
3. Way to much GC

The GC problem means that we are still creating too many objects. Can we do an audit and see if every one of these objects is needed?