@mhevery 2017-06-12
The issue is coming from `zoneAwareAddListener`

```TypeScript
    try {
      // In cross site contexts (such as WebDriver frameworks like Selenium),
      // accessing the handler object here will cause an exception to be thrown which
      // will fail tests prematurely.
      validZoneHandler = data.handler && data.handler.toString() === '[object FunctionWrapper]';
    } catch (error) {
      // we can still try to add the data.handler even we are in cross site context
      data.crossContext = true;
      return data.invokeAddFunc(addFnSymbol, data.handler);
    }
```

1. we have to remove the `try-catch` since that is preventing the code from optimizing.
2. Inside we do a check by calling data.handler.toString, this is expensive and has to go.