@JiaLiPassion 2017-06-25
@mhevery , I have updated the source.

1. remove old `addEventListener/removeEventListener` code.
2. make nodejs work in this new method.
And nodejs also get a 40% performance up.

- current version. 

new callback: 700ms 
reuse callback : 550ms

- improved version.

new callback: 440ms
reuse callback: 300ms

3. add new API for `browser`, now `EventTarget` support `eventListenerList` method to get all listeners for all event name or for specified event name

4. add new API for `browser` to support `removeAllListeners`

Please review, Thank you.